+++ jest.config.ts
@@ -1,14 +1,17 @@
 import type { Config } from '@jest/types';
 
-// Sync object
 const config: Config.InitialOptions = {
   preset: 'ts-jest',
   verbose: true,
-  automock: true,
-  transform: {
-    '^.+\\.(ts)?$': 'ts-jest',
-    '^.+\\.js$': 'babel-jest',
-  },
-  transformIgnorePatterns: ['/node_modules/(?!my-package)(.*)'],
+  automock: false,
+  clearMocks: true,
+  coveragePathIgnorePatterns: ['/node_modules/'],
+  coverageProvider: 'v8',
+  moduleDirectories: ['node_modules'],
+  moduleFileExtensions: ['js', 'ts'],
+  roots: ['tests'],
+  testMatch: ['**/__tests__/**/*.[jt]s?(x)', '**/?(*.)+(spec|test).[tj]s?(x)'],
+  testPathIgnorePatterns: ['/node_modules/'],
+  transformIgnorePatterns: ['/node_modules/', '\\.pnp\\.[^\\/]+$'],
 };
 export default config;
